WG/requirements.txt
blinker==1.9.0
bottle==0.13.2
branca==0.8.1
Cartopy==0.24.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.1.1
Flask==3.1.0
folium==0.19.5
fonttools==4.55.8
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proxy_tools==0.1.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
pyparsing==3.2.1
pyperclip==1.9.0
pyproj==3.7.1
pyserial==3.5
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywebview==5.4
PyYAML==6.0.2
ratelim==0.1.6
requests==2.32.3
scipy==1.15.2
serial==0.0.97
shapely==2.0.7
six==1.17.0
tk==0.1.0
tkintermapview==1.29
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
xyzservices==2025.1.0
